chore: add `utils.test.ts` file
